requirements.txt
Flask
PyPDF2
pillow
pytesseract
werkzeug